+++ straxen/__init__.py
@@ -1,10 +1,8 @@
-__version__ = '0.18.4'
+__version__ = '0.19.0'
 
-# Import of utilix config for only once. NB: Should be first due to circular imports
 from utilix import uconfig
-
-# Straxen imports:
 from .common import *
+# contexts.py below
 from .corrections_services import *
 from .get_corrections import *
 from .hitfinder_thresholds import *
@@ -17,10 +15,11 @@ from .online_monitor import *
 from .rundb import *
 from .scada import *
 from .bokeh_utils import *
+from .rucio import *
 
-# Nested structures:
 from . import plugins
 from .plugins import *
+
 from . import analyses
 
 # Do not make all contexts directly available under straxen.