+++ HISTORY.md
@@ -1,3 +1,47 @@
+0.19.0 / 2021-06-23
+-------------------
+minor changes
+- S1/S2 event patternfit and S1 AFT test (#499)
+- Change tight_coincidence (#564)
+- Fixing saturation correction bugs (#541)
+- Rewrite merge s2 (#548)
+- Compute width again after saturation correction (#542, #552)
+- Add rucio frontend (#472, #553)
+- Redo hit_thresholds (#543)
+- Standardize CMT options as (correction, version, nT=boolean) (#457, #532)
+
+patches and fixes:
+- z coordinate update (#535)
+- Fix example command (#547)
+- Don't import holoviews and ipywidgets (#551)
+- pre_apply_function from $HOME only in pytest (#559)
+- Rundb should not crash on fuzzy (#561)
+- Remove travis for testing from straxen (#557)
+- Fix missing info in bootstrax docs, fix #546 (#558)
+- Add scada interface to docs (#560)
+- Tweaks for new release 0.19.0 (#562)
+
+0.18.6-0.18.8 / 2021-06-03
+-------------------
+- Patches installation for pypi (#529, e880420, fce6d87)
+
+
+0.18.5 / 2021-06-03
+---------------------
+- Allow variable event duration (#494, #528)
+- Veto Proximity Plugin (#296)
+- Apply database function prior to returning the data (#497)
+- Max-size for rechunkable raw-records (#495)
+- Itp map patch (#471)
+- Bin updates (#502)
+- Split requirement files, set autoupdate dependabot (#504)
+- Fix failing tests (#503)
+- Reduce review dog verbosity (#498)
+- Reduce plugin testing time (#500)
+- Patch remap cabled (#501)
+- Fix veto veto regions (#493)
+
+
 0.18.4 / 2021-05-20
 ---------------------
 - Documentation and package maintenance (#481)